+++ changelog.md
@@ -1,10 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+### [2.3.0] -- 2025-05-01
+#### Added
+- Full support for MIRI/LRS observations.
+- Calculate covariance matrices in light curve fitting. 
+- Change outline of run_DMS and fit_lightcurves scripts to be more pythonic and avoid multiprocessing errors. 
+- Misc. MIRI-related updates to documentation.
+- Misc. bug fixes. 
+
 ### [2.2.0] -- 2025-04-14
 #### Added
-- Explicit support for NIRSpec PRISM observations.
-- Multiple updates to compatability with NIRSpec G395M as well as non-SUB2048 NIRSpec subarrays.
+- Explicit support for NIRSpec/PRISM observations.
+- Multiple updates to compatability with NIRSpec/G395M as well as non-SUB2048 NIRSpec subarrays.
 - Streamline spectral binning using the spectres package. 
 - Increment jwst version to 1.17.1.
 - Minor changes to format of Stage 3 data products (stellar spectra) such that first extension is the central wavelength of a given wavelength bin and second extension is the bin half-width.